app/requirements.txt

## Discord Req
discord.py
python-dotenv


## App Req
asyncpraw
praw
pathlib
